+++ src/tests/nist_x509.cpp
@@ -12,7 +12,6 @@
 #endif
 
 #include <algorithm>
-#include <iostream>
 #include <fstream>
 #include <iomanip>
 #include <string>
