+++ src/tests/test_aead.cpp
@@ -37,7 +37,7 @@ class AEAD_Tests : public Text_Based_Test
 
          if(!enc || !dec)
             {
-            warn_about_missing(algo);
+            result.note_missing(algo);
             return result;
             }
