+++ src/tests/test_hash.cpp
@@ -29,7 +29,7 @@ class Hash_Function_Tests : public Text_Based_Test
 
          if(providers.empty())
             {
-            warn_about_missing("block cipher " + algo);
+            result.note_missing("block cipher " + algo);
             return result;
             }
 
@@ -39,7 +39,7 @@ class Hash_Function_Tests : public Text_Based_Test
 
             if(!hash)
                {
-               warn_about_missing(algo + " from " + provider);
+               result.note_missing(algo + " from " + provider);
                continue;
                }
