+++ src/tests/test_mac.cpp
@@ -35,7 +35,7 @@ class Message_Auth_Tests : public Text_Based_Test
 
          if(providers.empty())
             {
-            warn_about_missing("block cipher " + algo);
+            result.note_missing("block cipher " + algo);
             return result;
             }
 
@@ -45,7 +45,7 @@ class Message_Auth_Tests : public Text_Based_Test
 
             if(!mac)
                {
-               warn_about_missing(algo + " from " + provider);
+               result.note_missing(algo + " from " + provider);
                continue;
                }
