+++ src/tests/test_stream.cpp
@@ -33,7 +33,7 @@ class Stream_Cipher_Tests : public Text_Based_Test
 
          if(providers.empty())
             {
-            warn_about_missing("block cipher " + algo);
+            result.note_missing("block cipher " + algo);
             return result;
             }
 
@@ -43,7 +43,7 @@ class Stream_Cipher_Tests : public Text_Based_Test
 
             if(!cipher)
                {
-               warn_about_missing(algo + " from " + provider);
+               result.note_missing(algo + " from " + provider);
                continue;
                }
